requirements.txt
kafka-python>=2.0.2
pandas>=2.0.0
requests>=2.31.0
matplotlib>=3.7.0
